+++ services/routine/internal/repository/act.go
@@ -1,24 +1,30 @@
 package repository
 
-import "routine/internal/domain/aggregates/act"
+import (
+	"context"
+	"routine/internal/domain/aggregates/act"
+)
 
 func NewActRepository(actRepo actRepository) *ActRepository {
-	return &ActRepository{actRepo}
+	return &ActRepository{repo: actRepo}
 }
 
 type actRepository interface {
-	Save(act *act.Act) (*act.Act, error)
-	FindByCode(code act.ActCode) (*act.Act, error)
+	Save(ctx context.Context, act *act.Act) (*act.Act, error)
+	FindByCode(ctx context.Context, code act.ActCode) (*act.Act, error)
+
+	Commit() error
+	Rollback(err error) error
 }
 
 type ActRepository struct {
-	actRepository
+	repo actRepository
 }
 
-func (ar *ActRepository) Save(act *act.Act) (*act.Act, error) {
-	return ar.actRepository.Save(act)
+func (ar *ActRepository) Save(ctx context.Context, act *act.Act) (*act.Act, error) {
+	return ar.repo.Save(ctx, act)
 }
 
-func (ar *ActRepository) FindByCode(code act.ActCode) (*act.Act, error) {
-	return ar.actRepository.FindByCode(code)
+func (ar *ActRepository) FindByCode(ctx context.Context, code act.ActCode) (*act.Act, error) {
+	return ar.repo.FindByCode(ctx, code)
 }
