+++ services/routine/internal/repository/error.go
@@ -8,4 +8,6 @@ func (e RepositoryError) Error() string {
 	return e.message
 }
 
+var ErrNoTransaction = RepositoryError{"no transaction"}
+
 var ErrNotFound = RepositoryError{"not found"}
